+++ src/util.h
@@ -18,7 +18,7 @@ bool IsPortInUse(int port);
 
 // string util
 std::string ToLower(std::string in);
-void BytesToHuman(char *s, unsigned long long n);
+void BytesToHuman(char *s, uint64_t n);
 std::string& Trim(std::string &in, std::string chars);
 void Split(std::string in, std::string delim, std::vector<std::string> *out);
